requirements.txt
beautifulsoup4==4.12.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.4.2
Pillow==9.4.0
pyquery==2.0.0
requests==2.28.1
streamlit==1.20.0
tld==0.13
validators==0.20.0
whois==0.9.27
